File structure modified throughout application and import updates
