+++ src/pages/Contact/Contact.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 const Contact = () => {
   return (
-    <div>
+    <main>
       Contact
-    </div>
+    </main>
   )
 };
